+++ inlang/source-code/paraglide/paraglide-js-adapter-next/examples/app/src/app/page.tsx
@@ -1,25 +1,24 @@
 import * as m from "@/paraglide/messages.js"
-import { languageTag } from "@/paraglide/runtime"
-import { ClientComponent } from "./ClientComponent"
-import { Link } from "@/lib/i18n"
 
 export default function Home() {
 	return (
-		<main>
-			<p>{m.greeting({ name: "Samuel", count: 5 })}</p>
-			<p>{m.currentLanguageTag({ languageTag: languageTag() })}</p>
-
-			<Link href="/about">{m.about()}</Link>
-			<br />
-
-			<Link href="/form">Form Flow</Link>
-			<br />
-
-			<Link href="/form" locale="en">
-				Form Flow (en)
-			</Link>
-			<br />
-			<ClientComponent />
-		</main>
+		<>
+			<h1>{m.paraglide_and_next_app_router()}</h1>
+			<p>{m.this_app_was_localised_with_paraglide()}</p>
+			<p>{m.switch_languages_in_top_right()}</p>
+			<p>{m.learn_more_at_following_links()}</p>
+			<ul style={{ color: "blue" }}>
+				<li>
+					<a href="https://inlang.com/m/osslbuzt/library-inlang-paraglideJsAdapterNextJs">
+						{m.paraglide_documentation()}
+					</a>
+				</li>
+				<li>
+					<a href="https://github.com/opral/monorepo/tree/main/inlang/source-code/paraglide/paraglide-js-adapter-next/examples/app">
+						{m.source_code()}
+					</a>
+				</li>
+			</ul>
+		</>
 	)
 }