+++ inlang/source-code/paraglide/paraglide-js-adapter-next/examples/app/src/app/redirect/page.tsx
@@ -1,5 +1,0 @@
-import { redirect } from "@/lib/i18n"
-
-export default async function Redirect() {
-	redirect("/about")
-}