+++ inlang/source-code/paraglide/paraglide-js-adapter-next/examples/app/src/lib/i18n.ts
@@ -2,15 +2,18 @@ import { AvailableLanguageTag } from "@/paraglide/runtime"
 import { createI18n } from "@inlang/paraglide-js-adapter-next"
 import * as m from "@/paraglide/messages"
 
-export const { Link, middleware, useRouter, usePathname, redirect, permanentRedirect } =
-	createI18n<AvailableLanguageTag>({
-		exclude: ["/not-translated"],
-		pathnames: {
-			"/about": m.about_path,
-			"/form": {
-				en: "/form",
-				de: "/formular",
-				"de-CH": "/formular",
-			},
-		},
-	})
+export const {
+	Link,
+	middleware,
+	useRouter,
+	usePathname,
+	redirect,
+	permanentRedirect,
+	localizePath,
+} = createI18n<AvailableLanguageTag>({
+	pathnames: {
+		"/about": m.about_path,
+	},
+
+	exclude: ["/not-translated"], //makes sure that the /not-translated page is not translated
+})