+++ inlang/source-code/paraglide/paraglide-js-adapter-next/examples/app/src/lib/ui/SelectLanguage.tsx
@@ -1,8 +1,14 @@
 "use client"
 import { availableLanguageTags, AvailableLanguageTag, languageTag } from "@/paraglide/runtime"
-import { useRouter, usePathname } from "./i18n"
+import { useRouter, usePathname } from "../i18n"
 
 export function SelectLanguage() {
+	const labels: Record<AvailableLanguageTag, string> = {
+		en: "🇬🇧 English",
+		de: "🇩🇪 Deutsch",
+		"de-CH": "🇨🇭 Schweizerdeutsch",
+	}
+
 	const pathname = usePathname()
 	const router = useRouter()
 
@@ -13,7 +19,7 @@ export function SelectLanguage() {
 		>
 			{availableLanguageTags.map((lang) => (
 				<option key={lang} value={lang}>
-					{lang}
+					{labels[lang]}
 				</option>
 			))}
 		</select>