+++ inlang/source-code/paraglide/paraglide-js-adapter-next/examples/pages/src/lib/Header.tsx
@@ -1,0 +1,26 @@
+import { LocaleSwitcher } from "./LocaleSwitcher"
+
+export function Header() {
+	return (
+		<header
+			style={{
+				display: "flex",
+				justifyContent: "space-between",
+				alignItems: "center",
+				padding: "1rem",
+				borderBottom: "1px solid #ccc",
+			}}
+		>
+			<nav
+				style={{
+					display: "flex",
+					gap: "1rem",
+				}}
+			>
+				{""}
+			</nav>
+
+			<LocaleSwitcher />
+		</header>
+	)
+}