+++ inlang/source-code/paraglide/paraglide-js-adapter-sveltekit/example/svelte.config.js
@@ -1,15 +1,9 @@
 import adapter from "@sveltejs/adapter-static"
 import { vitePreprocess } from "@sveltejs/vite-plugin-svelte"
-import { mdsvex } from "mdsvex"
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
-	preprocess: [
-		vitePreprocess(),
-		mdsvex({
-			extension: ".svx",
-		}),
-	],
+	preprocess: [vitePreprocess()],
 	kit: {
 		adapter: adapter(),
 		alias: {