+++ inlang/source-code/telemetry/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inlang/telemetry
 
+## 0.3.13
+
+### Patch Changes
+
+- @inlang/sdk@0.26.5
+
 ## 0.3.12
 
 ### Patch Changes